EX-220 - Made major enhancements to the ViewText and ViewBinary classes in the Stream library, along with other minor improvements to the stream library. Also fixed some build system issues and cleaned up some compiler warnings.

+++ Support Libraries/Stream/File.inl
@@ -117,7 +117,7 @@ bool File::Open(const std::wstring& filename, OpenMode openMode, CreateMode crea
 	{
 		bufferSize = abufferSize;
 		delete[] fileBuffer;
-		fileBuffer = new unsigned char[bufferSize];
+		fileBuffer = new unsigned char[(size_t)bufferSize];
 	}
 	bufferPosOffset = bufferSize;
 	bytesRemainingInBuffer = 0;
